Break split function into its own crate.
- Added tests
- Remove it from the delegation dir
- Added it as a crate in utils
