+++ beacon_chain/transition/src/delegation/mod.rs
@@ -1,3 +1,4 @@
+use super::honey_badger_split;
 use super::types;
 use super::TransitionError;
 use super::shuffling::shuffle;
